Fix application version check
The `NPPM_GETNPPVERSION` API returns a single DWORD that breaks down
to a high and low value (*). The low value is parsed from a version
string that is "descriptive" rather than semantic (**)

Most often there are 3 digits, e.g. "8.33". But with every "minor
version" bump, it's reduced to only 2 digits, e.g. "8.4"

So, plugins occasionally have to remember that 4 comes *after* 33 in
the (il)logic of N++'s versioning scheme

 ---
 (*) https://github.com/notepad-plus-plus/npp-usermanual/blob/master/content/docs/plugin-communication.md#nppm_getnppversion
 (**) https://github.com/notepad-plus-plus/notepad-plus-plus/blob/master/PowerEditor/src/resource.h#L24-L25