Make LCL modal dialogs behave more like VCL ones
`ShowModal` assumes the parent control is another component in the
framework's class hierarchy. In the case of one-off plugin dialogs, the
parent defaults to the global `Application` instance.

Delphi treats Notepad++ itself as the main application (similar to a
.NET plugin, which is no coincidence given .NET's development history).
So, for example, the main window is disabled when a modal form appears
and re-enabled when it returns.

Free Pascal's LCL doesn't behave as predictably. If the handle to the
main `Application` instance is the same as Notepad++, calling `Hide` on
the parent handle will *permanently* hide the program window, with no
way to focus it again, requiring the Task Manager to shut it down.

Not tampering with `Application`'s handle resolves most issues, but
creates a new one for modal forms. Since the main window remains enabled,
giving it focus sends the modal form to the back of the Z-order. A user
may forget the form is there and try to show it again, but the LCL throws
an exception if `ShowModal` is sent before a previous call has returned.

Note that, before and after this patch, it's possible to shut down the
application with an LCL modal form still showing, leaving a zombie
message loop that needs to be forcibly terminated.